CommentService 작성, 테스트코드 작성, 테스트코드 리펙토링 해야 함

+++ src/main/java/boardexample/myboard/domain/commnet/service/CommentServiceImpl.java
@@ -1,7 +1,20 @@
 package boardexample.myboard.domain.commnet.service;
 
 import boardexample.myboard.domain.commnet.Comment;
+import boardexample.myboard.domain.commnet.dto.CommentSaveDto;
+import boardexample.myboard.domain.commnet.dto.CommentUpdateDto;
+import boardexample.myboard.domain.commnet.exception.CommentException;
+import boardexample.myboard.domain.commnet.exception.CommentExceptionType;
 import boardexample.myboard.domain.commnet.repository.CommentRepository;
+import boardexample.myboard.domain.member.exception.MemberException;
+import boardexample.myboard.domain.member.exception.MemberExceptionType;
+import boardexample.myboard.domain.member.repository.MemberRepository;
+import boardexample.myboard.domain.post.Post;
+import boardexample.myboard.domain.post.exception.PostException;
+import boardexample.myboard.domain.post.exception.PostExceptionType;
+import boardexample.myboard.domain.post.repository.PostRepository;
+import boardexample.myboard.domain.post.service.PostService;
+import boardexample.myboard.global.util.security.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,31 +27,60 @@ import java.util.List;
 public class CommentServiceImpl implements CommentService{
 
     private final CommentRepository commentRepository;
-
+    private final MemberRepository memberRepository;
+    private final PostRepository postRepository;
 
     @Override
-    public void save(Comment comment) {
+    public void save(Long postId, CommentSaveDto commentSaveDto) {
+        Comment comment = commentSaveDto.toEntity();
+
+        comment.confirmWriter(memberRepository.findByUsername(SecurityUtil.getLoginUsername()).orElseThrow(() -> new MemberException(MemberExceptionType.NOT_FOUND_MEMBER)));
+
+        comment.confirmPost(postRepository.findById(postId).orElseThrow(() -> new PostException(PostExceptionType.POST_NOT_POUND)));
+
 
         commentRepository.save(comment);
+
     }
 
     @Override
-    @Transactional(readOnly = true)
-    public Comment findById(Long id) throws Exception {
-        return commentRepository.findById(id).orElseThrow(() -> new Exception("댓글이 없습니다."));
+    public void saveReComment(Long postId, Long parentId, CommentSaveDto commentSaveDto) {
+        Comment comment = commentSaveDto.toEntity();
+
+        comment.confirmWriter(memberRepository.findByUsername(SecurityUtil.getLoginUsername()).orElseThrow(() -> new MemberException(MemberExceptionType.NOT_FOUND_MEMBER)));
+
+        comment.confirmPost(postRepository.findById(postId).orElseThrow(() -> new PostException(PostExceptionType.POST_NOT_POUND)));
+
+        comment.confirmParent(commentRepository.findById(parentId).orElseThrow(() -> new CommentException(CommentExceptionType.NOT_POUND_COMMENT)));
+
+        commentRepository.save(comment);
+
     }
 
+
+
     @Override
-    @Transactional(readOnly = true)
-    public List<Comment> findAll() {
-        return commentRepository.findAll();
+    public void update(Long id, CommentUpdateDto commentUpdateDto) {
+
+        Comment comment = commentRepository.findById(id).orElseThrow(() -> new CommentException(CommentExceptionType.NOT_POUND_COMMENT));
+
+        if(!comment.getWriter().getUsername().equals(SecurityUtil.getLoginUsername())){
+            throw new CommentException(CommentExceptionType.NOT_AUTHORITY_UPDATE_COMMENT);
+        }
+
+        commentUpdateDto.content().ifPresent(comment::updateContent);
     }
 
 
 
     @Override
     public void remove(Long id) throws Exception {
-        Comment comment = commentRepository.findById(id).orElseThrow(() -> new Exception("댓글이 없습니다."));
+        Comment comment = commentRepository.findById(id).orElseThrow(() -> new CommentException(CommentExceptionType.NOT_POUND_COMMENT));
+
+        if(!comment.getWriter().getUsername().equals(SecurityUtil.getLoginUsername())){
+            throw new CommentException(CommentExceptionType.NOT_AUTHORITY_DELETE_COMMENT);
+        }
+
         comment.remove();
         List<Comment> removableCommentList = comment.findRemovableList();
         removableCommentList.forEach(removableComment -> commentRepository.delete(removableComment));
